requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
asttokens==3.0.1
beautifulsoup4==4.14.2
billiard==4.2.3
blis==1.3.3
catalogue==2.0.10
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.13
dateparser==1.2.2
decorator==5.2.1
Django==5.2.8
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.1
greenlet==3.2.4
idna==3.11
inflection==0.5.1
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
kombu==5.5.4
Markdown==3.10
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
MouseInfo==0.1.3
murmurhash==1.0.15
numpy==2.3.5
openapi-codec==1.3.2
packaging==25.0
parso==0.8.5
pexpect==4.9.0
playwright==1.56.0
playwright-stealth==2.0.0
preshed==3.0.12
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
simplejson==3.20.2
six==1.17.0
smart_open==7.5.0
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.2
stack-data==0.6.3
thinc==8.3.10
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
